Move metrics to transactron
