+++ transactron/lib/__init__.py
@@ -6,3 +6,4 @@ from .transformers import *  # noqa: F401
 from .reqres import *  # noqa: F401
 from .storage import *  # noqa: F401
 from .simultaneous import *  # noqa: F401
+from .metrics import *  # noqa: F401
